power: fix reset of low_power_trigger_level on boot
Previously, user-defined values for low_power_trigger_level below 20% were discarded and reset on boot due to a hardcoded minimum threshold.
This change ensures that the persisted user setting is respected and restores it early in the power management initialization.
Additionally, the minimum allowed trigger level is reduced from 20% to 10% to allow more flexibility for user-defined values.
Values like 10% or 15% are now preserved instead of being overwritten with the default.

Tested on LineageOS 22 (Android 15) with working persistency across reboots.

Change-Id: I3b17e8fedc2566a10779a6ce84c11b324d2bd6d4
Signed-off-by: Mario Suess <ms0281@gmx.de>